// api.ts
// Generated by https://quicktype.io

export interface API {
    copyright:       string;
    date:            string;
    explanation:     string;
    hdurl:           string;
    media_type:      string;
    service_version: string;
    title:           string;
    url:             string;
}

// Generated by https://quicktype.io

export interface SpaceCraft {
    count:    number;
    next:     string;
    previous: null;
    results:  Result[];
}

export interface Result {
    id:            number;
    url:           string;
    name:          string;
    agency:        Agency;
    in_use:        boolean;
    capability:    string;
    maiden_flight: string;
    human_rated:   boolean;
    crew_capacity: number | null;
    image_url:     string;
    nation_url:    null | string;
    wiki_link:     string;
    info_link:     string;
}

export interface Agency {
    id:            number;
    url:           string;
    name:          string;
    featured:      boolean;
    type:          null | string;
    country_code:  string;
    abbrev:        string;
    description:   string;
    administrator: null | string;
    founding_year: string;
    launchers:     string;
    spacecraft:    string;
    parent:        null | string;
    image_url:     null | string;
}